# push! performance tweaks